refactor: サジェスト取得をコンポーネント化、サジェストのソートを純関数に切り出し
